feat(hardware): :hammer: rpi ble app
BLE server for RPi with RDAC and pressure sensor

closes #253

+++ apps/rpi-ble/pressure/bmp3xx.py
@@ -2,10 +2,8 @@
 '''!
   based on @url  https://github.com/DFRobot/DFRobot_BMP3XX
 '''
-import sys
 import time
 
-import smbus
 import spidev
 import RPi.GPIO as GPIO
 
@@ -327,7 +325,7 @@ class BMP3XX(object):
         pd4 = pd3 + p11 * adc_p ** 3.0
         pressure = po1 + po2 + pd4
 
-        return round(pressure, -1), round(temperature, 0)
+        return int(round(pressure, -1)), int(round(temperature, 0))
 
     def _get_reg_temp_press_data(self):
         data = self._read_reg(BMP3XX_P_DATA_PA, 6)
